requirements.txt
django
djangorestframework
pillow
faiss-cpu
